+++ services/wallet-service/src/application/interface/dto/transaction/ProcessSplitPurchaseDTO.ts
@@ -1,7 +1,6 @@
 export interface ProcessSplitPurchaseDTO {
   buyerId: string;
   sellerId: string;
-  adminId: string;
   totalAmount: number;
   commissionAmount: number;
   artId: string;
