+++ services/wallet-service/src/application/usecases/wallet/ProcessSplitPurchaseUseCase.ts
@@ -4,6 +4,7 @@ import { ProcessSplitPurchaseDTO } from "../../interface/dto/transaction/Process
 import { TYPES } from "../../../infrastructure/inversify/types";
 import { IWalletRepository } from "../../../domain/repository/IWalletRepository";
 import { logger } from "../../../utils/logger";
+import { config } from "../../../infrastructure/config/env";
 
 @injectable()
 export class ProcessSplitPurchaseUseCase implements IProcessSplitPurchaseUseCase {
@@ -13,14 +14,15 @@ export class ProcessSplitPurchaseUseCase implements IProcessSplitPurchaseUseCase
   ) {}
 
   async execute(dto: ProcessSplitPurchaseDTO): Promise<boolean> {
+    const adminId = config.platform_admin_id
     logger.info(
-      `Processing split purchase for art ${dto.artId}: Buyer=${dto.buyerId}, Seller=${dto.sellerId}, Admin=${dto.adminId}, Total=${dto.totalAmount}, Commission=${dto.commissionAmount}`
+      `Processing split purchase for art ${dto.artId}: Buyer=${dto.buyerId}, Seller=${dto.sellerId}, Admin=${adminId}, Total=${dto.totalAmount}, Commission=${dto.commissionAmount}`
     );
 
     return this._walletRepository.processSplitPurchase(
       dto.buyerId,
       dto.sellerId,
-      dto.adminId,
+      adminId,
       dto.totalAmount,
       dto.commissionAmount,
       dto.artId
