+++ services/wallet-service/src/presentation/controllers/TransactionController.ts
@@ -134,57 +134,6 @@ export class TransactionController implements ITransactionController {
   };
 
 
-
-  //# ================================================================================================================
-  //# PROCESS SPLIT PURCHASE
-  //# ================================================================================================================
-  //# POST /api/v1/transaction/split-purchase
-  //# Request body: { buyerId, sellerId, adminId, totalAmount, commissionAmount, artId }
-  //# This controller handles art purchase transactions with commission split
-  //# ================================================================================================================
-  processSplitPurchase = async (
-    req: Request,
-    res: Response,
-    next: NextFunction
-  ): Promise<Response | void> => {
-    try {
-      const { buyerId, sellerId, adminId, totalAmount, commissionAmount, artId } = req.body;
-
-      logger.info(
-        `[TransactionController] Processing split purchase for art: ${artId} | buyer: ${buyerId} | seller: ${sellerId}`
-      );
-      
-      const dto: ProcessSplitPurchaseDTO = {
-        buyerId, 
-        sellerId, 
-        adminId, 
-        totalAmount, 
-        commissionAmount, 
-        artId
-      };
-
-      const success = await this._processSplitPurchaseUseCase.execute(dto);
-
-      if (success) {
-        logger.info(
-          `[TransactionController] Split purchase processed successfully for art: ${artId}`
-        );
-        return res
-          .status(HttpStatus.OK)
-          .json({ message: "Purchase successful" });
-      } else {
-        return res
-          .status(HttpStatus.BAD_REQUEST)
-          .json({ message: "Purchase failed" });
-      }
-    } catch (error) {
-      logger.error(
-        `[TransactionController] Error processing split purchase: ${error}`
-      );
-      next(error);
-    }
-  };
-
   //# ================================================================================================================
   //# PROCESS PAYMENT
   //# ================================================================================================================
