+++ services/wallet-service/src/presentation/interface/ITransactionController.ts
@@ -11,12 +11,6 @@ export interface ITransactionController {
     res: Response,
     next: NextFunction
   ) => Promise<Response | void>;
-
-  processSplitPurchase: (
-    req: Request,
-    res: Response,
-    next: NextFunction
-  ) => Promise<Response | void>;
   processPayment: (
     req: Request,
     res: Response,
